AmbientLight FPS control
